+++ spec/Knp/FriendlyContexts/Context/EntityContextSpec.php
@@ -3,8 +3,8 @@
 namespace spec\Knp\FriendlyContexts\Context;
 
 use Behat\Gherkin\Node\TableNode;
-use Doctrine\Common\Persistence\ManagerRegistry;
-use Doctrine\Common\Persistence\ObjectManager;
+use Doctrine\Persistence\ManagerRegistry;
+use Doctrine\Persistence\ObjectManager;
 use Doctrine\ORM\AbstractQuery;
 use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\Mapping\ClassMetadata;