+++ src/Knp/FriendlyContexts/Doctrine/ValidateObjectManager.php
@@ -1,0 +1,12 @@
+<?php
+
+
+namespace Knp\FriendlyContexts\Doctrine;
+
+use Doctrine\Common\Persistence\ObjectManager as CommonObjectManager;
+use Doctrine\Persistence\ObjectManager;
+
+
+trait ValidateObjectManager
+{
+}