+++ testapp/app/AppKernel.php
@@ -11,6 +11,7 @@ class AppKernel extends Kernel
             new Symfony\Bundle\FrameworkBundle\FrameworkBundle(),
             new Symfony\Bundle\TwigBundle\TwigBundle(),
             new Doctrine\Bundle\DoctrineBundle\DoctrineBundle(),
+            new FriendsOfBehat\SymfonyExtension\Bundle\FriendsOfBehatSymfonyExtensionBundle(),
             new Knp\FcTestBundle\KnpFcTestBundle(),
         );
 
@@ -19,11 +20,31 @@ class AppKernel extends Kernel
 
     public function registerContainerConfiguration(LoaderInterface $loader)
     {
-        $loader->load($this->getProjectDir() . '/app/config/config_' . $this->getEnvironment() . '.yml');
+        $configFile = 'config3-4.yml';
+        if (version_compare(self::VERSION, "5.0", ">=")) {
+            $configFile = 'config5.yml';
+        }
+        $loader->load($this->getProjectDir() . '/app/config/' . $configFile);
     }
 
     public function getProjectDir()
     {
         return realpath(__DIR__ . '/../');
+    }
+
+    /**
+     * Returns the kernel parameters.
+     *
+     * @return array An array of kernel parameters
+     */
+    protected function getKernelParameters()
+    {
+        $params = parent::getKernelParameters();
+
+        if ( ! isset($params['kernel.root_dir'])) {
+            $params['kernel.root_dir'] = $params['kernel.project_dir'] . DIRECTORY_SEPARATOR . 'app';
+        }
+
+        return $params;
     }
 }