+++ CHANGELOG.md
@@ -7,14 +7,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 
+### Changed
+- `ObjectManager` was dropped from the following method signatures:
+   - `Knp\FriendlyContexts\Doctrine\EntityHydrator::hydrate()`
+   - `Knp\FriendlyContexts\Doctrine\EntityHydrator::completeRequired()`
+   - `Knp\FriendlyContexts\Doctrine\EntityHydrator::completeFields()`
+   - `Knp\FriendlyContexts\Doctrine\EntityResolver::resolve()`
+   - `Knp\FriendlyContexts\Doctrine\EntityResolver::getClassesFromName()`
+   - `Knp\FriendlyContexts\Doctrine\EntityResolver::getMetadataFromProperty()`
+   - `Knp\FriendlyContexts\Doctrine\EntityResolver::getMetadataFromObject()`
+
 ### Removed
 - Dropped support for PHP 7.0
+- Dropped support for Symfony <3.4
 
 ## [0.10.0] - 2020-09-23
 ### Added
 - Add Support for PHP >=7.0,<7.5
 - Add Support for Alice v3
-- Add Support for Symfony 4
+- Add Support for Symfony 4 & 5
 
 ### Removed
 - Dropped support for PHP 5.6